Pembuatan Class Baru
Pembuatan Class Baru untuk data diri pegawai
