+++ DataDiri.java
@@ -1,0 +1,17 @@
+package Employee; 
+
+public class DataDiri{
+	private String firstName;
+	private String lastName;
+	private String idNumber;
+	`
+	public void setSpouse(String spouseName, String spouseIdNumber) {
+		this.spouseName = spouseName;
+		this.spouseIdNumber = idNumber;
+	}
+	
+	public void addChild(String childName, String childIdNumber) {
+		childNames.add(childName);
+		childIdNumbers.add(childIdNumber);
+	}
+}
